requirements.txt
python-telegram-bot==20.3
# Зависимости для логирования и отладки (опционально)
loguru>=0.6.0
